requirements.txt
asgiref==3.7.2
Django==4.2.2
django-jalali-date==1.0.2
djangorestframework==3.14.0
jdatetime==4.1.1
psycopg2==2.9.6
pytz==2023.3
sqlparse==0.4.4
tzdata==2023.3